// frontEnd/guess-the-song/src/app/Models/room.model.ts
export class Room {
  roomId?: number | undefined | null = null;
  joinedUsersIds: string | undefined | null = null;
  adminId: string | undefined | null = null;
  noOfQuestions: string | undefined | null = null;
  answerTime: string | undefined | null = null;
  noOfAnswers: string | undefined | null = null;
  roomCode: string | undefined | null = null;
}